+++ src/DnDSource.js
@@ -1,5 +1,5 @@
 import { DragSource } from 'react-dnd'
-import {DATETIME_FORMAT} from './index'
+import {ViewTypes, DATETIME_FORMAT} from './index'
 import {DnDTypes} from './DnDTypes'
 
 export default class DnDSource {
@@ -30,7 +30,14 @@ export default class DnDSource {
 
                 if(isEvent) {
                     const event = item;
-                    newStart = localeMoment(event.start).add(localeMoment(newStart).diff(localeMoment(initialStart)), 'ms').format(DATETIME_FORMAT);
+                    if(config.relativeMove) {
+                        newStart = localeMoment(event.start).add(localeMoment(newStart).diff(localeMoment(initialStart)), 'ms').format(DATETIME_FORMAT);
+                    } else {
+                        if(viewType !== ViewTypes.Day) {
+                            let tmpMoment = localeMoment(newStart);
+                            newStart = localeMoment(event.start).year(tmpMoment.year()).month(tmpMoment.month()).date(tmpMoment.date()).format(DATETIME_FORMAT);
+                        }
+                    }
                     newEnd = localeMoment(newStart).add(localeMoment(event.end).diff(localeMoment(event.start)), 'ms').format(DATETIME_FORMAT);
                 }